+++ packages/plugin-react-app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.2.1
+
+- [fix] bump version of `@builder/pack`(^0.6.0)
+- [fix] remove `webpackHotDevClient`
+- [fix] `vitePlugins` only take effect when `vite` is true
+
 ## 2.2.0
 
 - [feat] optimize runtime when build